pyproject.toml
[tool.poetry]
name = "tmi_archive"
version = "0.1.0"
description = ""
authors = ["simon <simon@simonnouwens.nl>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "==4.0.8"
django-reversion = "*"
django-reversion-compare = "*"
sentry-sdk = "*"
psycopg2-binary = "*"
boto3 = "==1.21.3"
django-storages = "*"
django-crum = "*"
django-dynamic-filenames = "*"
bootstrap4 = "*"
django-bootstrap4 = "*"
paramiko = "*"
django-crispy-forms = "*"
django-allauth = "*"
pandas = "*"
python-dotenv = "*"
django-tinymce = "*"
importlib-metadata = "*"
django-backblaze-b2 = "*"
typing-extensions = "*"
django-admin-sortable2 = "*"
django-select2 = "*"
"backports.zoneinfo" = "*"
"backports.weakref" = "*"
django-fullurl = "*"
djangorestframework = "*"
annotate = "*"
django-enum = "*"
mutagen = "*"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
